Implement DocumentRetrievalService and update RAGService alias; refactor imports and add corresponding tests
