+++ tapio/services/__init__.py
@@ -1,4 +1,4 @@
+from tapio.services.document_retrieval_service import RAGService
 from tapio.services.llm_service import LLMService
-from tapio.services.rag_service import RAGService
 
 __all__ = ["LLMService", "RAGService"]
